+++ server/start.go
@@ -192,7 +192,9 @@ func startInProcess(ctx *Context, cdc codec.JSONMarshaler, appCreator AppCreator
 		return err
 	}
 
-	if viper.GetBool("api.enable") {
+	config := config.GetConfig()
+	var apiSrv *api.Server
+	if config.API.Enable {
 		genDoc, err := genDocProvider()
 		if err != nil {
 			return err
@@ -208,19 +210,10 @@ func startInProcess(ctx *Context, cdc codec.JSONMarshaler, appCreator AppCreator
 			WithClient(local.New(tmNode)).
 			WithTrustNode(true)
 
-		apiSrv := api.New(ctx)
-		apiCfg := config.APIConfig{
-			Address:            viper.GetString("api.address"),
-			MaxOpenConnections: viper.GetUint("api.max-open-connections"),
-			RPCReadTimeout:     viper.GetUint("api.rpc-read-timeout"),
-			RPCWriteTimeout:    viper.GetUint("api.rpc-write-timeout"),
-			RPCMaxBodyBytes:    viper.GetUint("api.rpc-max-body-bytes"),
-			EnableUnsafeCORS:   viper.GetBool("api.enabled-unsafe-cors"),
-		}
-
+		apiSrv = api.New(ctx)
 		app.RegisterAPIRoutes(apiSrv)
 
-		if err := apiSrv.Start(apiCfg); err != nil {
+		if err := apiSrv.Start(config); err != nil {
 			return err
 		}
 	}
@@ -252,6 +245,10 @@ func startInProcess(ctx *Context, cdc codec.JSONMarshaler, appCreator AppCreator
 
 		if cpuProfileCleanup != nil {
 			cpuProfileCleanup()
+		}
+
+		if apiSrv != nil {
+			_ = apiSrv.Close()
 		}
 
 		ctx.Logger.Info("exiting...")