+++ tests/cli/simd_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/tests/cli"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/auth"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 )
 
@@ -99,7 +99,7 @@ func TestCLISimdAddGenesisAccount(t *testing.T) {
 
 	appCodec := f.EncodingConfig.Marshaler
 
-	accounts := auth.GetGenesisStateFromAppState(appCodec, genesisState).Accounts
+	accounts := authtypes.GetGenesisStateFromAppState(appCodec, genesisState).Accounts
 	balances := banktypes.GetGenesisStateFromAppState(f.Cdc, genesisState).Balances
 	balancesSet := make(map[string]sdk.Coins)