+++ x/auth/client/cli/broadcast.go
@@ -25,7 +25,7 @@ $ <appcli> tx broadcast ./mytxn.json
 `),
 		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx = clientCtx.Init()
+			clientCtx = clientCtx.Init().WithJSONMarshaler(cdc)
 
 			if clientCtx.Offline {
 				return errors.New("cannot broadcast tx during offline mode")